+++ juniper/src/tests/mod.rs
@@ -4,6 +4,8 @@ pub mod fixtures;
 #[cfg(test)]
 mod introspection_tests;
 #[cfg(test)]
+mod operation_not_supported;
+#[cfg(test)]
 mod query_tests;
 #[cfg(test)]
 mod schema_introspection;